Added a note about newer pip behaviour.
Since https://github.com/pypa/pip/issues/1423 it is no longer allowed to
install packages that are not hosted on PyPI itself by default. It can
be allowed on a per-package basis with the --allow-externals PACKAGE
option.
Since the latest argparse tarball comes from google code, it is
necessary to use --allow-externals for the argparse dependency.